pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multiview"
version = "0.1.0"
description = "A Python project"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "datasets>=2.0.0,<3.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "rank-bm25>=0.2.2",
    "google-genai>=1.57.0",
    "beautifulsoup4>=4.0.0",
    "requests>=2.0.0",
    "torch>=2.0.0",
    "transformers>=4.0.0",
    "numpy>=1.24.0",
    "accelerate>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=3.7.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
# API clients for inference providers
providers = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "google-genai>=0.2.0",
    "huggingface-hub>=0.20.0",
]
# Visualization tools
viz = [
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "jupyter>=1.0.0",
    "scikit-learn>=1.3.0",
    "umap-learn>=0.5.8",
    "numba>=0.60.0",
    "llvmlite>=0.44.0",
    "pillow>=10.0.0",
]
# All optional dependencies for full testing
all = [
    "multiview[dev,providers,viz]",
]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["multiview"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
]
markers = [
    "dev: lightweight sanity tests useful during development",
    "integration: end-to-end or pipeline-level tests",
    "external: tests that call external APIs / require network access",
    "needs_gpu: tests that require CUDA GPU for model inference",
    "needs_api: tests that require API keys and make API calls",
    "slow: tests that take a long time to run (e.g., downloading/loading large models)",
]
